+++ spec/scss_lint/cli_spec.rb
@@ -150,6 +150,18 @@ describe SCSSLint::CLI do
       end
     end
 
+    context 'when specified SCSS file globs match no files' do
+      before do
+        SCSSLint::FileFinder.any_instance.stub(:find)
+          .and_raise(SCSSLint::Exceptions::NoFilesError)
+      end
+
+      it 'exits with an appropriate status code' do
+        subject.should_receive(:halt).with(:no_files)
+        safe_run
+      end
+    end
+
     context 'when all specified SCSS files are filtered by exclusions' do
       before do
         SCSSLint::FileFinder.any_instance.stub(:find)